refactor(memory): split episodic.py into focused modules (RFC 0005, PR 8b)
Extract NoteStore + Note to notes.py and migrations/scoring to
migrations.py. EpisodicMemory delegates note operations to NoteStore
via composition. Pure move refactoring — no logic changes.
