+++ FILEMAP.md
@@ -1,14 +1,14 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-04-14  
-> **188** tracked files
+> **190** tracked files
 
 ## Summary
 
 | Top-level directory | Files |
 |---------------------|-------|
 | `internal/` | 40 |
-| `agents/` | 33 |
+| `agents/` | 35 |
 | `tests/` | 28 |
 | `docs/` | 21 |
 | `(root)` | 19 |
@@ -26,7 +26,7 @@
 
 | Extension | Count |
 |-----------|-------|
-| `.py` | 70 |
+| `.py` | 72 |
 | `.md` | 35 |
 | `.go` | 35 |
 | `.yaml` | 22 |
@@ -91,6 +91,8 @@
 │   ├── memory/
 │   │   ├── __init__.py
 │   │   ├── episodic.py
+│   │   ├── migrations.py
+│   │   ├── notes.py
 │   │   ├── relationship.py
 │   │   └── working.py
 │   ├── persona.py
