+++ agents/memory/__init__.py
@@ -2,7 +2,8 @@
 
 from typing import Protocol, runtime_checkable
 
-from .episodic import Episode, EpisodicMemory, Note
+from .episodic import Episode, EpisodicMemory
+from .notes import Note, NoteStore
 from .relationship import Interaction, RelationshipMemory, RelationshipSummary
 from .working import ContextSection, WorkingMemory, estimate_tokens
 
@@ -35,6 +36,7 @@ __all__ = [
     "Interaction",
     "MemoryLifecycle",
     "Note",
+    "NoteStore",
     "RelationshipMemory",
     "RelationshipSummary",
     "WorkingMemory",
