+++ agents/memory/episodic.py
@@ -5,7 +5,8 @@ Stores summaries of conversations, decisions, and outcomes in SQLite
 with FTS5 full-text search for relevance-ranked retrieval.
 
 Also provides agent-initiated note storage (migration v2) for
-structured knowledge the agent chooses to persist.
+structured knowledge the agent chooses to persist, delegated to
+:class:`~agents.memory.notes.NoteStore`.
 """
 
 from __future__ import annotations
@@ -15,10 +16,20 @@ import logging
 import sqlite3
 import time
 import uuid
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any
 
 import aiosqlite
+
+from .migrations import (
+    _FTS5_DDL,
+    _NOTES_FTS5_DDL,
+    _SCORE_EXPR,
+    _SCORE_EXPR_BARE,
+    _apply_migrations,
+    _fts5_available,
+)
+from .notes import Note, NoteStore
 
 if TYPE_CHECKING:
     from ..llm_client import LLMClient
@@ -47,30 +58,6 @@ class Episode:
     compression_level: int  # 0=raw, 1=summarized, 2=distilled
 
 
-@dataclass
-class Note:
-    """An agent-initiated note persisted via memory tools."""
-
-    id: str
-    agent_id: str
-    topic: str
-    content: str
-    tags: list[str] = field(default_factory=list)
-    access_count: int = 0
-    created_at: float = 0.0
-    updated_at: float = 0.0
-
-
-# Maximum content size for a single note (10 KB).
-_MAX_NOTE_CONTENT_BYTES = 10_240
-
-# Column list for SELECT queries on the notes table.
-_NOTE_COLS = (
-    "id", "agent_id", "topic", "content", "tags_json",
-    "access_count", "created_at", "updated_at",
-)
-_NOTE_SELECT = ", ".join(_NOTE_COLS)
-
 # Column list for SELECT queries — keeps _row_to_episode() positional
 # mapping stable when future migrations add columns to the episodes table.
 _EPISODE_COLS = (
@@ -85,221 +72,6 @@ _EPISODE_SELECT_ALIASED = ", ".join(f"e.{c}" for c in _EPISODE_COLS)
 # result sets and resource exhaustion.
 _MAX_RECALL_LIMIT = 100
 
-# ─── Shared scoring SQL fragments ──────────────────────────
-
-# Non-BM25 scoring components shared across _recall_fts5(), _recall_like(),
-# and _recall_recency().  Extracted to avoid maintaining the same formula in
-# three SQL strings (F-3a-2).
-#
-# importance is wrapped with (0.1 + importance * 0.9) so that episodes with
-# importance=0.0 still receive a non-zero score (10% baseline) instead of
-# being invisible in ranked recall (F-3a-1).
-#
-# A single template parameterizes the column prefix ("e." for JOINed queries,
-# "" for bare queries) so formula tuning is a single edit (F-59-2).
-#
-# Security: {p} is ONLY filled with hardcoded column prefixes ("e." or "").
-# NEVER pass user or LLM input into _SCORE_TEMPLATE.format().
-# All dynamic values use parameterized ? placeholders.
-_SCORE_TEMPLATE = (
-    "(0.1 + {p}importance * 0.9)"
-    " * (1.0 + ln(1 + {p}access_count))"
-    " * (1.0 / (1 + (? - {p}created_at) / 86400.0))"  # ? = time.time()
-)
-_SCORE_EXPR = _SCORE_TEMPLATE.format(p="e.")
-_SCORE_EXPR_BARE = _SCORE_TEMPLATE.format(p="")
-
-
-# ─── Schema migrations ─────────────────────────────────────
-
-# Forward-only migrations: (version, description, SQL).
-# Each migration's SQL may contain multiple statements separated by ";".
-MIGRATIONS: list[tuple[int, str, str]] = [
-    (
-        1,
-        "Initial schema: episodes + agent_state + FTS5",
-        """
-        CREATE TABLE IF NOT EXISTS episodes (
-            id TEXT PRIMARY KEY,
-            agent_id TEXT NOT NULL,
-            summary TEXT NOT NULL,
-            context_json TEXT,
-            outcome TEXT,
-            importance REAL DEFAULT 0.5,
-            access_count INTEGER DEFAULT 0,
-            last_accessed_at REAL,
-            tags_json TEXT,
-            created_at REAL NOT NULL,
-            compressed_at REAL,
-            compression_level INTEGER DEFAULT 0
-        );
-
-        CREATE INDEX IF NOT EXISTS idx_episodes_agent
-            ON episodes(agent_id);
-        CREATE INDEX IF NOT EXISTS idx_episodes_importance
-            ON episodes(importance DESC);
-        CREATE INDEX IF NOT EXISTS idx_episodes_created
-            ON episodes(created_at DESC);
-
-        CREATE TABLE IF NOT EXISTS agent_state (
-            agent_id TEXT PRIMARY KEY,
-            interaction_count INTEGER DEFAULT 0,
-            persona_state_json TEXT,
-            updated_at REAL NOT NULL
-        );
-        """,
-    ),
-    (
-        2,
-        "Notes table, FTS5 index, and sync triggers",
-        """
-        CREATE TABLE IF NOT EXISTS notes (
-            id TEXT PRIMARY KEY,
-            agent_id TEXT NOT NULL,
-            topic TEXT NOT NULL,
-            content TEXT NOT NULL,
-            tags_json TEXT,
-            access_count INTEGER DEFAULT 0,
-            created_at REAL NOT NULL,
-            updated_at REAL NOT NULL
-        );
-
-        CREATE INDEX IF NOT EXISTS idx_notes_agent
-            ON notes(agent_id);
-        CREATE INDEX IF NOT EXISTS idx_notes_topic
-            ON notes(agent_id, topic);
-        CREATE INDEX IF NOT EXISTS idx_notes_created
-            ON notes(created_at DESC);
-        """,
-    ),
-    (
-        3,
-        "Relationships and interactions tables",
-        """
-        CREATE TABLE IF NOT EXISTS relationships (
-            agent_id TEXT NOT NULL,
-            other_agent_id TEXT NOT NULL,
-            trust_score REAL DEFAULT 0.5,
-            interaction_count INTEGER DEFAULT 0,
-            last_interaction_at REAL,
-            notes TEXT,
-            PRIMARY KEY (agent_id, other_agent_id)
-        );
-
-        CREATE TABLE IF NOT EXISTS interactions (
-            id TEXT PRIMARY KEY,
-            agent_id TEXT NOT NULL,
-            other_agent_id TEXT NOT NULL,
-            interaction_type TEXT NOT NULL,
-            outcome TEXT,
-            sentiment REAL DEFAULT 0.0,
-            created_at REAL NOT NULL
-        );
-
-        CREATE INDEX IF NOT EXISTS idx_relationships_agent
-            ON relationships(agent_id);
-        -- Composite covering index for get_relationship_summary() query:
-        -- WHERE agent_id=? AND other_agent_id=? ORDER BY created_at DESC LIMIT N
-        -- Replaces separate agent and created_at indexes; the composite
-        -- index satisfies both the WHERE filter and ORDER BY in a single
-        -- index scan, avoiding a temp sort.
-        CREATE INDEX IF NOT EXISTS idx_interactions_lookup
-            ON interactions(agent_id, other_agent_id, created_at DESC);
-        """,
-    ),
-]
-
-# FTS5 DDL — applied only when FTS5 is available.
-_FTS5_DDL = """
-CREATE VIRTUAL TABLE IF NOT EXISTS episodes_fts USING fts5(
-    summary, context_json,
-    content=episodes, content_rowid=rowid
-);
-
-CREATE TRIGGER IF NOT EXISTS episodes_ai AFTER INSERT ON episodes BEGIN
-    INSERT INTO episodes_fts(rowid, summary, context_json)
-        VALUES (new.rowid, new.summary, new.context_json);
-END;
-
-CREATE TRIGGER IF NOT EXISTS episodes_ad AFTER DELETE ON episodes BEGIN
-    INSERT INTO episodes_fts(episodes_fts, rowid, summary, context_json)
-        VALUES ('delete', old.rowid, old.summary, old.context_json);
-END;
-
-CREATE TRIGGER IF NOT EXISTS episodes_au AFTER UPDATE ON episodes BEGIN
-    INSERT INTO episodes_fts(episodes_fts, rowid, summary, context_json)
-        VALUES ('delete', old.rowid, old.summary, old.context_json);
-    INSERT INTO episodes_fts(rowid, summary, context_json)
-        VALUES (new.rowid, new.summary, new.context_json);
-END;
-"""
-
-# FTS5 DDL for notes — applied only when FTS5 is available.
-_NOTES_FTS5_DDL = """
-CREATE VIRTUAL TABLE IF NOT EXISTS notes_fts USING fts5(
-    topic, content, tags_json,
-    content=notes, content_rowid=rowid
-);
-
-CREATE TRIGGER IF NOT EXISTS notes_ai AFTER INSERT ON notes BEGIN
-    INSERT INTO notes_fts(rowid, topic, content, tags_json)
-        VALUES (new.rowid, new.topic, new.content, new.tags_json);
-END;
-
-CREATE TRIGGER IF NOT EXISTS notes_ad AFTER DELETE ON notes BEGIN
-    INSERT INTO notes_fts(notes_fts, rowid, topic, content, tags_json)
-        VALUES ('delete', old.rowid, old.topic, old.content, old.tags_json);
-END;
-
-CREATE TRIGGER IF NOT EXISTS notes_au AFTER UPDATE ON notes BEGIN
-    INSERT INTO notes_fts(notes_fts, rowid, topic, content, tags_json)
-        VALUES ('delete', old.rowid, old.topic, old.content, old.tags_json);
-    INSERT INTO notes_fts(rowid, topic, content, tags_json)
-        VALUES (new.rowid, new.topic, new.content, new.tags_json);
-END;
-"""
-
-
-async def _apply_migrations(db: aiosqlite.Connection) -> None:
-    """Apply all pending schema migrations."""
-    await db.execute(
-        "CREATE TABLE IF NOT EXISTS schema_version "
-        "(version INTEGER PRIMARY KEY, applied_at REAL NOT NULL, description TEXT)"
-    )
-    async with db.execute("SELECT MAX(version) FROM schema_version") as cursor:
-        row = await cursor.fetchone()
-    current = (row[0] if row and row[0] is not None else 0)
-
-    for version, desc, sql in MIGRATIONS:
-        if version > current:
-            # NOTE: executescript() implicitly calls COMMIT before executing,
-            # so the DDL and the version record below are NOT atomic.  If the
-            # process crashes between executescript() and the INSERT, the
-            # migration is applied but not recorded — causing a re-run on
-            # restart.  This is safe for v1 because all statements use
-            # IF NOT EXISTS guards.  Future non-idempotent migrations (ALTER
-            # TABLE, data transforms) MUST use individual db.execute() calls
-            # inside a manually managed transaction instead.
-            await db.executescript(sql)
-            await db.execute(
-                "INSERT INTO schema_version VALUES (?, ?, ?)",
-                (version, time.time(), desc),
-            )
-            logger.info("Applied migration v%d: %s", version, desc)
-    await db.commit()
-
-
-async def _fts5_available(db: aiosqlite.Connection) -> bool:
-    """Test FTS5 availability with a throwaway virtual table."""
-    try:
-        await db.execute(
-            "CREATE VIRTUAL TABLE IF NOT EXISTS _fts5_test USING fts5(x)"
-        )
-        await db.execute("DROP TABLE IF EXISTS _fts5_test")
-        return True
-    except Exception:
-        return False
-
 
 # ─── EpisodicMemory ────────────────────────────────────────
 
@@ -312,6 +84,7 @@ class EpisodicMemory:
         self._db_path = db_path
         self._db: aiosqlite.Connection | None = None
         self._fts5: bool = False
+        self._note_store: NoteStore | None = None
 
     @property
     def agent_id(self) -> str:
@@ -352,18 +125,28 @@ class EpisodicMemory:
                 "Install a Python build with FTS5 support for production use."
             )
 
+        self._note_store = NoteStore(
+            agent_id=self._agent_id, db=self._db, fts5=self._fts5,
+        )
+
     async def close(self) -> None:
         """Close the database connection."""
         if self._db is not None:
             await self._db.close()
             self._db = None
+            self._note_store = None
 
     def _ensure_db(self) -> aiosqlite.Connection:
         if self._db is None:
             raise RuntimeError("EpisodicMemory not initialized — call initialize() first")
         return self._db
 
-    # ─── CRUD ───────────────────────────────────────────────
+    def _ensure_note_store(self) -> NoteStore:
+        if self._note_store is None:
+            raise RuntimeError("EpisodicMemory not initialized — call initialize() first")
+        return self._note_store
+
+    # ─── Episode CRUD ──────────────────────────────────────
 
     async def store_episode(
         self,
@@ -749,7 +532,7 @@ class EpisodicMemory:
             )
         return deleted
 
-    # ─── Notes CRUD ─────────────────────────────────────────
+    # ─── Notes (delegated to NoteStore) ────────────────────
 
     async def store_note(
         self,
@@ -762,48 +545,9 @@ class EpisodicMemory:
 
         Returns the generated note ID.
         """
-        db = self._ensure_db()
-        # Validate max_notes: _prune_notes() computes
-        # LIMIT MAX(0, count - max_notes + 1), so max_notes=0 would
-        # delete ALL existing notes.  Reject at the public API boundary
-        # even though create_memory_tools() always passes 500 (F-59-1).
-        if max_notes < 1:
-            raise ValueError(f"max_notes must be >= 1, got {max_notes}")
-        if not topic or not topic.strip():
-            raise ValueError("topic must not be empty")
-        if not content or not content.strip():
-            raise ValueError("content must not be empty")
-        content_bytes = content.encode("utf-8")
-        if len(content_bytes) > _MAX_NOTE_CONTENT_BYTES:
-            raise ValueError(
-                f"content exceeds {_MAX_NOTE_CONTENT_BYTES} byte limit "
-                f"({len(content_bytes)} bytes)"
-            )
-
-        # Prune oldest low-access notes if at capacity.
-        await self._prune_notes(db, max_notes)
-
-        note_id = str(uuid.uuid4())
-        now = time.time()
-        await db.execute(
-            """
-            INSERT INTO notes
-                (id, agent_id, topic, content, tags_json,
-                 access_count, created_at, updated_at)
-            VALUES (?, ?, ?, ?, ?, 0, ?, ?)
-            """,
-            (
-                note_id,
-                self._agent_id,
-                topic.strip(),
-                content,
-                json.dumps(tags or []),
-                now,
-                now,
-            ),
+        return await self._ensure_note_store().store_note(
+            topic, content, tags=tags, max_notes=max_notes,
         )
-        await db.commit()
-        return note_id
 
     async def recall_notes(
         self,
@@ -815,176 +559,19 @@ class EpisodicMemory:
 
         Increments access_count on returned notes.
         """
-        if limit < 1:
-            raise ValueError(f"limit must be >= 1, got {limit}")
-        limit = min(limit, _MAX_RECALL_LIMIT)
-        db = self._ensure_db()
-
-        if query and self._fts5:
-            rows = await self._recall_notes_fts5(db, query, limit)
-        elif query:
-            rows = await self._recall_notes_like(db, query, limit)
-        else:
-            rows = await self._recall_notes_recency(db, limit)
-
-        notes = [self._row_to_note(row) for row in rows]
-
-        if notes:
-            ids = [n.id for n in notes]
-            placeholders = ",".join("?" for _ in ids)
-            await db.execute(
-                f"UPDATE notes SET access_count = access_count + 1 "
-                f"WHERE id IN ({placeholders})",
-                ids,
-            )
-            await db.commit()
-            for note in notes:
-                note.access_count += 1
-
-        return notes
+        return await self._ensure_note_store().recall_notes(query, limit=limit)
 
     async def update_note(self, note_id: str, content: str) -> bool:
         """Update note content. Topic and tags preserved. Returns True if found."""
-        db = self._ensure_db()
-        if not content or not content.strip():
-            raise ValueError("content must not be empty")
-        content_bytes = content.encode("utf-8")
-        if len(content_bytes) > _MAX_NOTE_CONTENT_BYTES:
-            raise ValueError(
-                f"content exceeds {_MAX_NOTE_CONTENT_BYTES} byte limit "
-                f"({len(content_bytes)} bytes)"
-            )
-        now = time.time()
-        cursor = await db.execute(
-            "UPDATE notes SET content = ?, updated_at = ? "
-            "WHERE id = ? AND agent_id = ?",
-            (content, now, note_id, self._agent_id),
-        )
-        await db.commit()
-        return cursor.rowcount > 0
+        return await self._ensure_note_store().update_note(note_id, content)
 
     async def delete_note(self, note_id: str) -> bool:
         """Delete a note by ID (agent-scoped). Returns True if found."""
-        db = self._ensure_db()
-        cursor = await db.execute(
-            "DELETE FROM notes WHERE id = ? AND agent_id = ?",
-            (note_id, self._agent_id),
-        )
-        await db.commit()
-        return cursor.rowcount > 0
+        return await self._ensure_note_store().delete_note(note_id)
 
     async def count_notes(self) -> int:
         """Return the number of notes for this agent."""
-        db = self._ensure_db()
-        async with db.execute(
-            "SELECT COUNT(*) FROM notes WHERE agent_id = ?",
-            (self._agent_id,),
-        ) as cursor:
-            row = await cursor.fetchone()
-        return row[0] if row else 0
-
-    async def _prune_notes(self, db: aiosqlite.Connection, max_notes: int) -> None:
-        """Remove oldest low-access notes when count >= max_notes.
-
-        Uses a single atomic DELETE with a subquery to avoid a TOCTOU race
-        between SELECT count and DELETE that could occur when multiple
-        EpisodicMemory instances share a DB file (F-3b-1).
-        """
-        await db.execute(
-            """
-            DELETE FROM notes WHERE agent_id = ? AND id IN (
-                SELECT id FROM notes
-                WHERE agent_id = ?
-                ORDER BY access_count ASC, created_at ASC
-                LIMIT MAX(0, (SELECT COUNT(*) FROM notes WHERE agent_id = ?) - ? + 1)
-            )
-            """,
-            (self._agent_id, self._agent_id, self._agent_id, max_notes),
-        )
-
-    async def _recall_notes_fts5(
-        self,
-        db: aiosqlite.Connection,
-        query: str,
-        limit: int,
-    ) -> list[aiosqlite.Row]:
-        """FTS5 search across topic, content, and tags."""
-        try:
-            async with db.execute(
-                f"""
-                SELECT {", ".join(f"n.{c}" for c in _NOTE_COLS)}
-                FROM notes_fts fts
-                JOIN notes n ON n.rowid = fts.rowid
-                WHERE notes_fts MATCH ?
-                  AND n.agent_id = ?
-                ORDER BY fts.rank * -1 DESC
-                LIMIT ?
-                """,
-                (query, self._agent_id, limit),
-            ) as cursor:
-                return await cursor.fetchall()
-        except sqlite3.OperationalError as exc:
-            logger.warning(
-                "Notes FTS5 query failed for %r, falling back to LIKE: %s",
-                query,
-                exc,
-            )
-            return await self._recall_notes_like(db, query, limit)
-
-    async def _recall_notes_like(
-        self,
-        db: aiosqlite.Connection,
-        query: str,
-        limit: int,
-    ) -> list[aiosqlite.Row]:
-        """LIKE fallback when FTS5 is unavailable."""
-        escaped = query.replace("\\", "\\\\").replace("%", "\\%").replace("_", "\\_")
-        pattern = f"%{escaped}%"
-        async with db.execute(
-            f"""
-            SELECT {_NOTE_SELECT}
-            FROM notes
-            WHERE agent_id = ?
-              AND (topic LIKE ? ESCAPE '\\'
-                   OR content LIKE ? ESCAPE '\\'
-                   OR tags_json LIKE ? ESCAPE '\\')
-            ORDER BY updated_at DESC
-            LIMIT ?
-            """,
-            (self._agent_id, pattern, pattern, pattern, limit),
-        ) as cursor:
-            return await cursor.fetchall()
-
-    async def _recall_notes_recency(
-        self,
-        db: aiosqlite.Connection,
-        limit: int,
-    ) -> list[aiosqlite.Row]:
-        """No query — return most recently updated notes."""
-        async with db.execute(
-            f"""
-            SELECT {_NOTE_SELECT}
-            FROM notes
-            WHERE agent_id = ?
-            ORDER BY updated_at DESC
-            LIMIT ?
-            """,
-            (self._agent_id, limit),
-        ) as cursor:
-            return await cursor.fetchall()
-
-    def _row_to_note(self, row: aiosqlite.Row) -> Note:
-        """Convert a database row to a Note dataclass."""
-        return Note(
-            id=row[0],
-            agent_id=row[1],
-            topic=row[2],
-            content=row[3],
-            tags=json.loads(row[4]) if row[4] else [],
-            access_count=row[5],
-            created_at=row[6],
-            updated_at=row[7],
-        )
+        return await self._ensure_note_store().count_notes()
 
     # ─── Interaction counter (auto_reflect_after) ───────────
